requirements.txt
Flask==2.3.7
Pillow==10.1.0
Werkzeug==2.3.7
gunicorn==21.2.0
